Add DATA.BAS sample.
The DATA.BAS sample explores reading/writing Applesoft pointers in the
zero page area to allow RESTORing to a particular place in the code
rather than at the beginning.  This would allow the opcode table to be
read in once and then multiple passes made over the source code in my
assembler.

I also updated APACHE.BAS to place it at the end of the source file so
that the linear line searches conducted by GOTO and GOSUBs don't have
to keep skipping over it.  It will also show up at the end of LISTing
so that the user won't have to pause the listing to see it.