First version of SASM.BAS which can assemble 1 line
The simple assembler is not complete enough that it can successfully
assemble one line of 6502 assembly language code.  Yay!!

I have also added the latest map and pseudo code which include changes
that I made as I was completing this version of the assembler and
getting it to successfully assemble its first line.