Cleanup DATA statements for opcode tables.
I removed spaces and quotation marks that weren't required from the
DATA statements used for the 6502 instruction and directive tables.
This will save some space in the BASIC code.

I also verified that the opcode table in the Applesoft BASIC code
matches the opcode table that I have in my snap cross assembler.  At
this point, I think the assembler is ready for me to actually try
using it to write, assemble, and run some 6502 code.