Commit first version of SAMPLE.BAS
This is the start of a 65c02 assembly language sample that I am working
on using my Applesoft BASIC based assembler.  So far it is just able to
prompt the user for a filename using Apple II monitor calls.

So far the assembler is working well.  If there were things that could
improve the experience, they would be:
* Further improve performance.
* Supporting operators like '+' and '-' in expressions would be nice.
  One area that I have had to work around this lack of support in this
  sample can be found with the code around PRINTF1 and PRINTF2.  I am
  actually just modifying the absolute address for a LDX ????,X
  instruction.  With expressions I could have just used a LDX
  instruction in the code with a fake absolute address and a single
  label and then add 1 and 2 to the opcode address to modify the
  address bytes.
* ASC could support double quotes to specify negative ASCII text which
  is used by the monitor APIs.