SAMPLE.BAS opens/closes ProDOS files
I updated SAMPLE.BAS so that it can now open the specified ProDOS file.
I want to test it over the ADTPro virtual drive to see if it works to
open files on a virtual HDD.  If that works, I will start reading the
contents of the file and the fun can really start.