Small performance update to assembler.
Moved extractNextField() up to the top of the BASIC code to start at
line 10.  This was done because it was a hot function that showed up in
my profiling runs.  Applesoft uses a linear search to resolve
GOSUB targets so placing this hot function towards the beginning of
the code improves performance.  I also inlined the findEndOfString(),
findNextWhitespace(), and findNextNonWhitespace() routines which are
only used from within extractNextField.  These changes got me a 5%
improvement in run time.  Initializing the I, K, A%, A$, T$, and R%
variables early in the initVariables() routine got me another 1%
improvement in runtime.  This is due to the fact that these variables
are used from within this hot extractNextField() routine and they too
are resolved via a linear search.