Update sample
Move data runs so that they are top aligned in a page.  This makes the
future write to disk path easier since it will just need to increment
an index register and check for overflow to 0.  Having the runs in
memory so that they don't cross page boundaries also keeps the timing
consistent.